requirements.txt
google-generativeai
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2
python-dateutil
requests
readability-lxml
python-dotenv
fastapi
uvicorn[standard]
pydantic
fastapi-cors
python-multipart